+++ CHANGELOG.md
@@ -30,6 +30,7 @@ contains albums and artists and song now has nullable fields and `artists` and
 - Add control message `remove-playlist` that can remove playlist at the given
   index.
 - Add option to negate matching modes with `^`.
+- Support negation in filter with `^`.
 
 ### Changes
 - Add new id control messages `InsertIntoPlaylist` and `RemoveFromPlaylist`.